requirements.txt
mysql-connector-python
ppretty
prettytable
protobuf
wcwidth
